pyproject.toml
[project]
name = "polyagent"
version = "0.1.0"
description = "Multi-agent framework for code generation, API testing, and UI automation"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
dependencies = [
    "python-dotenv>=1.0.0",
    "httpx[socks]>=0.28.1",
    "langchain-core>=1.2.23",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "playwright>=1.58.0",
    "rich>=14.3.3",
    "tavily-python>=0.7.23",
    "typer>=0.24.1",
    "readchar>=4.2.2",
    "requests>=2.33.0",
]

[project.scripts]
polyagent = "cli.main:main_cli"

[tool.hatch.build.targets.wheel]
packages = ["agents", "graphs", "tools", "prompts", "schema", "cli"]

[tool.hatch.build.targets.wheel.force-include]
"prompts/system_prompts" = "prompts/system_prompts"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
